requirements.txt
tweepy
kivy
kivymd
numba
numpy
click
pillow